Realisation of renderer to cpp files
