+++ src/native/renderer/Operable.cpp
@@ -1,0 +1,5 @@
+#pragma once
+
+class Operable {
+    virtual void clear();
+};
