+++ src/native/renderer/headers/multiversion/client/graphics/Operation.hpp
@@ -24,55 +24,29 @@ namespace lagrangian {
 
         template<class T> class VertexTransformationOperation : public VertexOperation, public Transformation<Vertex, T> {
             public:
-            virtual void apply(Vertex& v) override {
-                apply(v);
-            }
-            virtual void operate(std::vector<Operable>& vs) override {
-                std::vector<Operable>::iterator it;
-                for(it = vs.begin(); it != vs.end(); it++) {
-                    apply(*dynamic_cast<Vertex*>(&*it));
-                }
-            }
+            virtual void apply(Vertex& v) override;
+            virtual void operate(std::vector<Operable>& vs) override;
         };
 
         class UVOperation : public VertexTransformationOperation<Vector2<long>> {
             public:
-            UVOperation(UVOperation& op) : UVOperation(op.defaultval) {}
-            UVOperation() : UVOperation(0, 0) {}
-            UVOperation(Vector2<long> val) {
-                defaultval = val;
-            }
-            UVOperation(long x, long y) {
-                defaultval = {.x=x, .y=y};
-            }
-            virtual void apply(Vertex& v) {
-                apply(v.uv);
-            }
+            UVOperation(UVOperation& op) : UVOperation(op.defaultval);
+            UVOperation() : UVOperation(0, 0);
+            UVOperation(Vector2<long> val);
+            UVOperation(long x, long y);
+            virtual void apply(Vertex& v);
         };
 
         class UVScaleOperation : public UVOperation {
             public:
-            UVScaleOperation() : UVOperation() { }
-            UVScaleOperation(Vector2<long> val) : UVOperation(val) { }
-            UVScaleOperation(long x, long y) : UVOperation(x, y) { }
+            UVScaleOperation() : UVOperation();
+            UVScaleOperation(Vector2<long> val) : UVOperation(val);
+            UVScaleOperation(long x, long y) : UVOperation(x, y);
 
-            virtual Transformation<Vertex, Vector2<long>>& inverse() {
-                UVScaleOperation u(1/defaultval.x, 1/defaultval.y);
-                
-                static UVScaleOperation& us = u;
-                return us;
-            };
+            virtual Transformation<Vertex, Vector2<long>>& inverse();
+            virtual void apply(Vector2<long>& v) override;
 
-            virtual void apply(Vector2<long>& v) override {
-                v.x *= defaultval.x;
-                v.y *= defaultval.y;
-            }
-
-            virtual void applyInv(Vector2<long>& v) {
-                v.x *= -defaultval.x;
-                v.y *= -defaultval.y;
-            }
-
+            virtual void applyInv(Vector2<long>& v);
         };
     }
 }
