+++ src/native/renderer/headers/multiversion/client/graphics/Vertex.hpp
@@ -11,22 +11,9 @@ namespace lagrangian {
             Vector2<long> uv;
             Colour colour = Colour::NONE;
 
-            Vertex(Vector3<long> pos, Vector2<long> uv, Colour colour) {
-                this->pos = pos;
-                this->uv = uv;
-                this->colour = colour;
-            }
+            Vertex(Vector3<long> pos, Vector2<long> uv, Colour colour);
 
-            void clear() {
-				this->pos.x = 0;
-                this->pos.y = 0;
-                this->pos.z = 0;
-
-                this->uv.x = 0;
-                this->uv.y = 0;
-
-                this->colour = Colour::NONE;
-			}
+            void clear();
         };
     }
 }
