+++ src/native/utils/headers/Localization.hpp
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <dirent.h>
 
 #include <string>
@@ -26,8 +27,8 @@ class HashedString {
 class Localization {
 public:
 	char filler[4];
-	std::__ndk1::string fullLanguageCode; // 4 - getFullLanguageCode from 1.13.0.6
-	std::__ndk1::map<std::__ndk1::string, std::__ndk1::string> strings; // 8 - _getStrings from 1.13.0.6
+	std::__ndk1::string fullLanguageCode;
+	std::__ndk1::map<std::__ndk1::string, std::__ndk1::string> strings;
 	std::__ndk1::map<std::__ndk1::string, std::__ndk1::string>& _getStrings() const;
     std::__ndk1::string get(std::__ndk1::string&, std::vector<std::__ndk1::string> const&) const;
     std::__ndk1::string _getSimple(std::__ndk1::string) const;
